Import of PCOLLINS/HTTP-DAV-0.29 from CPAN.
gitpan-cpan-distribution: HTTP-DAV
gitpan-cpan-version:      0.29
gitpan-cpan-path:         PCOLLINS/HTTP-DAV-0.29.tar.gz
gitpan-cpan-author:       PCOLLINS
gitpan-cpan-maturity:     released

+++ t/5_proppatch.t
@@ -3,56 +3,23 @@ use strict;
 use HTTP::DAV;
 use Test;
 use Cwd;
+use lib 't';
+use TestDetails qw($test_user $test_pass $test_url $test_cwd do_test fail_tests test_callback);
 
 # Tests basic proppatch.
 
 my $TESTS;
-BEGIN {
-    require "t/TestDetails.pm"; import TestDetails;
-    $TESTS=14;
-    plan tests => $TESTS
-}
+$TESTS=14;
+plan tests => $TESTS;
+fail_tests($TESTS) unless $test_url =~ /http/;
 
-# I'll skip all the tests if you haven't set a default url
-if ( TestDetails::url() !~ /http/ ) {
-   print  "You need to set a test url in the t/TestDetails.pm module.\n";
-   for(1..$TESTS) { skip(1,1); }
-   exit;
-}
 
-my $user = TestDetails::user();
-my $pass = TestDetails::pass();
-my $url = TestDetails::url();
-my $cwd = getcwd(); # Remember where we started.
+my $user = $test_user;
+my $pass = $test_pass;
+my $url = $test_url;
+my $cwd = $test_cwd; # Remember where we started
 
 HTTP::DAV::DebugLevel(3);
-
-######################################################################
-# UTILITY FUNCTION: 
-#    do_test <op_result>, <expected_result>, <message>
-# IT was getting tedious doing the error handling so 
-# I built this little routine, Makes the test cases easier to read.
-sub do_test {
-   my($dav,$result,$expected,$message) = @_;
-   $expected = 1 if !defined $expected;
-   my $ok;
-   my $davmsg = $dav->message;
-   if ($expected) {
-      if ( $ok = ok($result,$expected) ) {
-         print "$message succeeded\n";
-      } else {
-         print "$message failed: \"$davmsg\"\n";
-      }
-   } else {
-      if ( $ok = ok($result,$expected) ) {
-         print "$message failed (as expected): \"$davmsg\"\n";
-      } else {
-         print "$message succeeded (unexpectedly): \"$davmsg\"\n";
-      }
-   }
-   return $ok;
-}
-######################################################################
 
 # Make a directory with our process id after it 
 # so that it is somewhat random
@@ -68,11 +35,20 @@ Setup.
    OPEN
    MKCOL perldav_test
    CWD perldav_test
+   PUT perldav_test/file.txt
+
+   #is option(perldav_test,PROPFIND)
+   #is option(perldav_test/file.txt, PROPFIND)
 
 Test 1. We want to test a set prop sequence.
-   PROPPATCH (set patrick:test_prop=test_val)
+   if is_option(perldav_test,PROPFIND) {
+      PROPPATCH(perldav_test, set patrick:test_prop=test_val)
+   }
+   if is_option(perldav_test/file.txt,PROPFIND) {
+      PROPPATCH(perldav_test/file.txt, set patrick:test_prop=test_val)
+   }
 
-Test 2. Then an remove prop sequence
+Test 2. Then a remove prop sequence
    PROPPATCH perldav_test (remove patrick:test_prop)
 
 Test 3. Then lock perldav_test and do a proppatch. No namespace
@@ -87,36 +63,67 @@ Test 3. Then lock perldav_test and do a proppatch. No namespace
 my $dav1 = HTTP::DAV->new();
 $dav1->credentials( $user, $pass, $url );
 do_test $dav1, $dav1->open  ( $url ),  1,"OPEN $url";
+
+# Determine server's willingness to proppatching and locking
+# IIS5 currently does not support pp on files or colls.
+my $options =$dav1->options();
+my $coll_proppatch=( $options=~/\bPROPPATCH\b/)?1:0;
+my $coll_lock=     ( $options=~/\bLOCK\b/     )?1:0;
+my $cps = ($coll_proppatch)?"supports":"does not support";
+my $cls = ($coll_lock     )?"supports":"does not support";
+print "$options\n";
+print "** Server $cps proppatch against collections ** \n";
+print "** Server $cls locking against collections ** \n";
+
+if (!$coll_proppatch) {
+   skip_num($TESTS-1); # We've already done one test on the open
+   exit;
+}
+
+
+######################################################################
+my $resource;
+
 do_test $dav1, $dav1->mkcol ($newdir), 1,"MKCOL $newdir";
 do_test $dav1, $dav1->cwd   ($newdir), 1,"CWD $newdir";
 
-my $resource = $dav1->get_workingresource();
 
-# Test 1
-my $resp = $resource->proppatch(-namespace=>'patrick',
-                     -propname=>'test_prop',
-                     -propvalue=>'test_val'
-                    );
-do_test $resp, $resp->is_success, 1,"proppatch set test_prop";
-$resp = $resource->propfind(-depth=>0);
-if ($resp->is_success) {
-   do_test $resp, $resource->get_property('test_prop'),'test_val',"propfind get_property test_prop";
+## Test 1.
+do_test $dav1, 
+   $dav1->proppatch(-namespace=>'patrick',
+                    -propname=>'test_prop',
+                    -propvalue=>'test_val'),
+   '/Resource/', 
+   "proppatch set test_prop";
+
+$resource = $dav1->propfind(-depth=>0);
+if ($resource) {
+   do_test $dav1, 
+           $resource->get_property('test_prop'),
+           'test_val',
+           "propfind get_property test_prop";
 } else {
    print "Couldn't perform propfind\n";
    ok 0;
 }
 print $resource->as_string;
 
-# Test 2
-$resp = $resource->proppatch(-namespace=>'patrick',
-                     -propname=>'test_prop',
-                     -action=>'remove'
-                    );
-do_test $resp, $resp->is_success, 1,"proppatch remove test_prop";
-print $resp->as_string;
-$resp = $resource->propfind(-depth=>0);
-if ($resp->is_success) {
-   do_test $resp, $resource->get_property('test_prop'),'',"propfind get_property test_prop";
+
+## Test 2
+do_test 
+   $dav1, 
+   $dav1->proppatch(-namespace=>'patrick',
+                    -propname=>'test_prop',
+                    -action=>'remove'),
+   '/Resource/', 
+   "proppatch remove test_prop";
+
+$resource = $dav1->propfind(-depth=>0);
+if ($resource) {
+   do_test $dav1, 
+           $resource->get_property('test_prop'),
+           '',
+           "propfind get_property test_prop";
 } else {
    print "Couldn't perform propfind\n";
    ok 0;
@@ -124,43 +131,53 @@ if ($resp->is_success) {
 print $resource->as_string;
 
 ######################################################################
-# Test 3a
-do_test $dav1, $dav1->lock  (),          1,"LOCK";
+if ($coll_lock) {
+   do_test $dav1, $dav1->lock(),          1,"LOCK";
+   
+   # Test 3a
+   do_test 
+      $dav1, 
+      $dav1->set_prop(-propname=>'test_prop',-propvalue=>'test_value2'),
+      '/Resource/', 
+      "proppatch set DAV:test_prop";
+   
+   $resource = $dav1->propfind(-depth=>0);
+   if ($resource) {
+      do_test $dav1, 
+              $resource->get_property('test_prop'),
+              'test_value2',
+              "propset get_property DAV:test_prop";
+   } else {
+      print "Couldn't perform propfind\n";
+      ok 0;
+   }
+   print $resource->as_string;
 
-$resp = $resource->proppatch(
-                     -propname=>'test_prop',
-                     -propvalue=>'test_value2'
-                    );
+   # Test 3b
+   do_test 
+      $dav1, 
+      $dav1->unset_prop(-propname=>'test_prop',-namespace=>'DAV'),
+      '/Resource/', 
+      "unset_prop DAV:test_prop";
+   
+   $resource = $dav1->propfind(-depth=>0);
+   if ($resource) {
+      do_test $dav1, 
+              $resource->get_property('test_prop'),
+              '',
+              "propfind get_property DAV:test_prop";
+   } else {
+      print "Couldn't perform propfind\n";
+      ok 0;
+   }
+   print $resource->as_string;
 
-do_test $resp, $resp->is_success, 1,"proppatch set DAV:test_prop";
-
-$resp = $resource->propfind(-depth=>0);
-
-if ($resp->is_success) {
-   do_test $resp, $resource->get_property('test_prop'),'test_value2',"propfind get_property DAV:test_prop";
-} else {
-   print "Couldn't perform propfind\n";
-   ok 0;
+   do_test $dav1, $dav1->unlock(),          1,"UNLOCK";
 }
-
-# Test 3b
-$resp = $resource->proppatch(-namespace=>'DAV',
-                     -propname=>'test_prop',
-                     -action=>'remove'
-                    );
-
-do_test $resp, $resp->is_success, 1,"proppatch remove DAV:test_prop";
-$resp = $resource->propfind(-depth=>0);
-if ($resp->is_success) {
-   do_test $resp, $resource->get_property('test_prop'),'',"propfind get_property test_prop";
-} else {
-   print "Couldn't perform propfind\n";
-   ok 0;
-}
-
-print $resource->as_string;
-do_test $dav1, $dav1->unlock(),          1,"UNLOCK";
 
 # Cleanup
-$dav1->cwd("..");
-do_test $dav1, $dav1->delete($newdir),      1,"DELETE $newdir";
+if ( $test_url =~ /http/ ) {
+   print "Cleaning up\n";
+   $dav1->cwd("..");
+   do_test $dav1, $dav1->delete($newdir),      1,"DELETE $newdir";
+}
